requirements.txt
scipy>=1.1.0
numpy>=1.14.0
tensorflow_gpu>=1.4.0
tqdm>=4.0.0
